feat: Customizing the app
Files and components renaming.

Code styles updates.

Integrations test update.

Resolves: #1
Resolves: #3
