+++ polymer.json
@@ -1,9 +1,7 @@
 {
   "entrypoint": "index.html",
-  "shell": "src/components/my-app.js",
-  "sources": [
-    "images/**/*"
-  ],
+  "shell": "src/components/resume-app.js",
+  "sources": ["images/**/*"],
   "extraDependencies": [
     "manifest.json",
     "node_modules/@webcomponents/webcomponentsjs/**",
@@ -12,10 +10,7 @@
   "builds": [
     {
       "name": "esm-bundled",
-      "browserCapabilities": [
-        "es2015",
-        "modules"
-      ],
+      "browserCapabilities": ["es2015", "modules"],
       "js": {
         "minify": true
       },
@@ -30,9 +25,7 @@
     },
     {
       "name": "es6-bundled",
-      "browserCapabilities": [
-        "es2015"
-      ],
+      "browserCapabilities": ["es2015"],
       "js": {
         "compile": "es2015",
         "minify": true,
