+++ push-manifest.json
@@ -1,10 +1,10 @@
 {
   "/": {
-    "src/components/my-app.js": {
+    "src/components/resume-app.js": {
       "type": "script",
       "weight": 1
     },
-    "src/components/my-view1.js": {
+    "src/components/resume-view1.js": {
       "type": "script",
       "weight": 1
     },
@@ -14,11 +14,11 @@
     }
   },
   "/view1": {
-    "src/components/my-app.js": {
+    "src/components/resume-app.js": {
       "type": "script",
       "weight": 1
     },
-    "src/components/my-view1.js": {
+    "src/components/resume-view1.js": {
       "type": "script",
       "weight": 1
     },
@@ -28,11 +28,11 @@
     }
   },
   "/view2": {
-    "src/components/my-app.js": {
+    "src/components/resume-app.js": {
       "type": "script",
       "weight": 1
     },
-    "src/components/my-view2.js": {
+    "src/components/resume-view2.js": {
       "type": "script",
       "weight": 1
     },
@@ -42,11 +42,11 @@
     }
   },
   "/view3": {
-    "src/components/my-app.js": {
+    "src/components/resume-app.js": {
       "type": "script",
       "weight": 1
     },
-    "src/components/my-view3.js": {
+    "src/components/resume-view3.js": {
       "type": "script",
       "weight": 1
     },
