+++ src/actions/app.ts
@@ -1,13 +1,3 @@
-/**
-@license
-Copyright (c) 2018 The Polymer Project Authors. All rights reserved.
-This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
-The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
-The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
-Code distributed by Google as part of the polymer project is also
-subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
-*/
-
 import { Action, ActionCreator } from 'redux';
 import { ThunkAction } from 'redux-thunk';
 import { RootState } from '../store.js';
@@ -17,16 +7,27 @@ export const UPDATE_DRAWER_STATE = 'UPDATE_DRAWER_STATE';
 export const OPEN_SNACKBAR = 'OPEN_SNACKBAR';
 export const CLOSE_SNACKBAR = 'CLOSE_SNACKBAR';
 
-export interface AppActionUpdatePage extends Action<'UPDATE_PAGE'> {page: string};
-export interface AppActionUpdateOffline extends Action<'UPDATE_OFFLINE'> {offline: boolean};
-export interface AppActionUpdateDrawerState extends Action<'UPDATE_DRAWER_STATE'> {opened: boolean};
-export interface AppActionOpenSnackbar extends Action<'OPEN_SNACKBAR'> {};
-export interface AppActionCloseSnackbar extends Action<'CLOSE_SNACKBAR'> {};
-export type AppAction = AppActionUpdatePage | AppActionUpdateOffline | AppActionUpdateDrawerState | AppActionOpenSnackbar | AppActionCloseSnackbar;
+export interface AppActionUpdatePage extends Action<'UPDATE_PAGE'> {
+  page: string;
+}
+export interface AppActionUpdateOffline extends Action<'UPDATE_OFFLINE'> {
+  offline: boolean;
+}
+export interface AppActionUpdateDrawerState extends Action<'UPDATE_DRAWER_STATE'> {
+  opened: boolean;
+}
+export interface AppActionOpenSnackbar extends Action<'OPEN_SNACKBAR'> {}
+export interface AppActionCloseSnackbar extends Action<'CLOSE_SNACKBAR'> {}
+export type AppAction =
+  | AppActionUpdatePage
+  | AppActionUpdateOffline
+  | AppActionUpdateDrawerState
+  | AppActionOpenSnackbar
+  | AppActionCloseSnackbar;
 
 type ThunkResult = ThunkAction<void, RootState, undefined, AppAction>;
 
-export const navigate: ActionCreator<ThunkResult> = (path: string) => (dispatch) => {
+export const navigate: ActionCreator<ThunkResult> = (path: string) => dispatch => {
   // Extract the page name from path.
   const page = path === '/' ? 'view1' : path.slice(1);
 
@@ -38,23 +39,23 @@ export const navigate: ActionCreator<ThunkResult> = (path: string) => (dispatch)
   dispatch(updateDrawerState(false));
 };
 
-const loadPage: ActionCreator<ThunkResult> = (page: string) => (dispatch) => {
-  switch(page) {
+const loadPage: ActionCreator<ThunkResult> = (page: string) => dispatch => {
+  switch (page) {
     case 'view1':
-      import('../components/my-view1.js').then(() => {
+      import('../components/resume-view1.js').then(() => {
         // Put code in here that you want to run every time when
-        // navigating to view1 after my-view1.js is loaded.
+        // navigating to view1 after resume-view1.js is loaded.
       });
       break;
     case 'view2':
-      import('../components/my-view2.js');
+      import('../components/resume-view2.js');
       break;
     case 'view3':
-      import('../components/my-view3.js');
+      import('../components/resume-view3.js');
       break;
     default:
       page = 'view404';
-      import('../components/my-view404.js');
+      import('../components/resume-view404.js');
   }
 
   dispatch(updatePage(page));
@@ -69,16 +70,18 @@ const updatePage: ActionCreator<AppActionUpdatePage> = (page: string) => {
 
 let snackbarTimer: number;
 
-export const showSnackbar: ActionCreator<ThunkResult> = () => (dispatch) => {
+export const showSnackbar: ActionCreator<ThunkResult> = () => dispatch => {
   dispatch({
     type: OPEN_SNACKBAR
   });
   window.clearTimeout(snackbarTimer);
-  snackbarTimer = window.setTimeout(() =>
-    dispatch({ type: CLOSE_SNACKBAR }), 3000);
+  snackbarTimer = window.setTimeout(() => dispatch({ type: CLOSE_SNACKBAR }), 3000);
 };
 
-export const updateOffline: ActionCreator<ThunkResult> = (offline: boolean) => (dispatch, getState) => {
+export const updateOffline: ActionCreator<ThunkResult> = (offline: boolean) => (
+  dispatch,
+  getState
+) => {
   // Show the snackbar only if offline status changes.
   if (offline !== getState().app!.offline) {
     dispatch(showSnackbar());
