--- src/components/my-view404.ts
@@ -1,31 +1,0 @@
-/**
-@license
-Copyright (c) 2018 The Polymer Project Authors. All rights reserved.
-This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
-The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
-The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
-Code distributed by Google as part of the polymer project is also
-subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
-*/
-
-import { html } from '@polymer/lit-element';
-import { PageViewElement } from './page-view-element.js';
-
-// These are the shared styles needed by this element.
-import { SharedStyles } from './shared-styles.js';
-
-class MyView404 extends PageViewElement {
-  protected render() {
-    return html`
-      ${SharedStyles}
-      <section>
-        <h2>Oops! You hit a 404</h2>
-        <p>The page you're looking for doesn't seem to exist. Head back
-           <a href="/">home</a> and try again?
-        </p>
-      </section>
-    `
-  }
-}
-
-window.customElements.define('my-view404', MyView404);
