+++ src/components/shop-products.ts
@@ -21,7 +21,7 @@ import './shop-item.js';
 import { getAllProducts, addToCart } from '../actions/shop.js';
 
 // These are the elements needed by this element.
-import { addToCartIcon } from './my-icons.js';
+import { addToCartIcon } from './resume-icons.js';
 
 // These are the shared styles needed by this element.
 import { ButtonSharedStyles } from './button-shared-styles.js';
