+++ test/integration/router.js
@@ -32,7 +32,7 @@ describe('routing tests', function() {
 
   it('the page selector switches pages', async function() {
     await page.goto(`${appUrl}`);
-    await page.waitForSelector('my-app', {visible: true});
+    await page.waitForSelector('resume-app', {visible: true});
 
     await testNavigation(page, 'view2', 'View Two');
     await testNavigation(page, 'view3', 'View Three');
@@ -41,7 +41,7 @@ describe('routing tests', function() {
 
   it('the page selector switches pages in a different way', async function() {
     await page.goto(`${appUrl}`);
-    await page.waitForSelector('my-app', {visible: true});
+    await page.waitForSelector('resume-app', {visible: true});
 
     // Setup
     await page.evaluate(() => {
@@ -78,7 +78,7 @@ async function testNavigation(page, href, linkText) {
   const shadowSelector = `a[href="/${href}"]`;
 
   // Does the link say the right thing?
-  const myApp = await page.$('my-app');
+  const myApp = await page.$('resume-app');
   const myText = await page.evaluate(getShadowRootChildProp, myApp, selector, 'textContent');
   expect(await myText).equal(linkText);
 
@@ -89,7 +89,7 @@ async function testNavigation(page, href, linkText) {
 }
 
 async function testNavigationInADifferentWay(page, href, linkText) {
-  const query = `my-app::shadow a[href="/${href}"]`;
+  const query = `resume-app::shadow a[href="/${href}"]`;
 
   const linkHandle = await page.evaluateHandle((query) => window.deepQuerySelector(query), query);
   const text = await page.evaluate((el) => el.textContent, linkHandle);
