+++ test/unit/index.html
@@ -1,30 +1,30 @@
 <!--
-@license
-Copyright (c) 2018 The Polymer Project Authors. All rights reserved.
-This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
-The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
-The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
-Code distributed by Google as part of the polymer project is also
-subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
+  @license
+  Copyright (c) 2018 The Polymer Project Authors. All rights reserved.
+  This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
+  The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
+  The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
+  Code distributed by Google as part of the polymer project is also
+  subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
 -->
 
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
-    <meta charset="utf-8">
+    <meta charset="utf-8" />
     <title>Tests</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
     <script src="../../node_modules/wct-browser-legacy/browser.js"></script>
   </head>
   <body>
     <script>
       WCT.loadSuites([
-        // Load 'my-view1.html' test suite, using native shadow dom
+        // Load 'resume-view1.html' test suite, using native shadow dom
         'counter-element.html',
-        // Load 'my-view1.html' test suite, using shadydom
+        // Load 'resume-view1.html' test suite, using shadydom
         'counter-element.html?wc-shadydom=true&wc-ce=true',
         'views-a11y.html',
-        'views-a11y.html?wc-shadydom=true&wc-ce=true',
+        'views-a11y.html?wc-shadydom=true&wc-ce=true'
       ]);
     </script>
   </body>
