+++ test/unit/views-a11y.html
@@ -1,19 +1,19 @@
 <!--
-@license
-Copyright (c) 2018 The Polymer Project Authors. All rights reserved.
-This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
-The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
-The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
-Code distributed by Google as part of the polymer project is also
-subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
+  @license
+  Copyright (c) 2018 The Polymer Project Authors. All rights reserved.
+  This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
+  The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
+  The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
+  Code distributed by Google as part of the polymer project is also
+  subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
 -->
 
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
   <head>
-    <meta charset="utf-8">
+    <meta charset="utf-8" />
     <title>views a11y</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
 
     <script>
       // Redux assumes `process.env.NODE_ENV` exists in the ES module build.
@@ -25,54 +25,54 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
     <script src="../../node_modules/wct-browser-legacy/browser.js"></script>
 
     <!-- Import the element to test -->
-    <script type="module" src="../../src/components/my-view1.js"></script>
-    <script type="module" src="../../src/components/my-view2.js"></script>
-    <script type="module" src="../../src/components/my-view3.js"></script>
-    <script type="module" src="../../src/components/my-view404.js"></script>
+    <script type="module" src="../../src/components/resume-view1.js"></script>
+    <script type="module" src="../../src/components/resume-view2.js"></script>
+    <script type="module" src="../../src/components/resume-view3.js"></script>
+    <script type="module" src="../../src/components/resume-view404.js"></script>
   </head>
   <body>
     <test-fixture id="view1">
       <template>
-        <my-view1 active></my-view1>
+        <resume-view1 active></resume-view1>
       </template>
     </test-fixture>
     <test-fixture id="view2">
       <template>
-        <my-view2 active></my-view2>
+        <resume-view2 active></resume-view2>
       </template>
     </test-fixture>
     <test-fixture id="view3">
       <template>
-        <my-view3 active></my-view3>
+        <resume-view3 active></resume-view3>
       </template>
     </test-fixture>
     <test-fixture id="view404">
       <template>
-        <my-view404 active></my-view404>
+        <resume-view404 active></resume-view404>
       </template>
     </test-fixture>
 
     <script type="module">
       import 'axe-core/axe.min.js';
-      import {axeReport} from 'pwa-helpers/axe-report.js';
+      import { axeReport } from 'pwa-helpers/axe-report.js';
 
       suite('views a11y tests', function() {
-        test('my-view1', function() {
+        test('resume-view1', function() {
           const el = fixture('view1');
           return axeReport(el);
         });
 
-        test('my-view2', function() {
+        test('resume-view2', function() {
           const el = fixture('view2');
           return axeReport(el);
         });
 
-        test('my-view3', function() {
+        test('resume-view3', function() {
           const el = fixture('view3');
           return axeReport(el);
         });
 
-        test('my-view404', function() {
+        test('resume-view404', function() {
           const el = fixture('view404');
           return axeReport(el);
         });
